HeaderForm 생성, 반응형 작업
